+++ front/src/features/todos/components/TodoCompleteBtn.tsx
@@ -16,7 +16,7 @@ export function TodoCompleteBtn({ id }: Props) {
     try {
       // dispatch
       await dispatch(updateTodoIsCompleted(id)).unwrap();
-      toastSuccessRB('Todoを更新しました', { autoClose: 2000 })
+      toastSuccessRB('Todoを完了しました', { autoClose: 2000 })
     } catch {
       const errorMessage = error || 'Todoの更新に失敗しました'
       toastErrorRB(errorMessage, { autoClose: 2000 })
@@ -29,3 +29,4 @@ export function TodoCompleteBtn({ id }: Props) {
     </>
   )
 }
+
